+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.2.0
 
-- Add support to Dart v3.0.0 (#52)[https://github.com/PostHog/posthog-flutter/pull/52]
+- Add support to Dart v3.0.0 [#52](https://github.com/PostHog/posthog-flutter/pull/52)
 
 ## 3.1.0