+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- Migrate to Java 8 and minSdkVersion 19 [#54](https://github.com/PostHog/posthog-flutter/pull/54)
+
 ## 3.2.0
 
 - Add support to Dart v3.0.0 [#52](https://github.com/PostHog/posthog-flutter/pull/52)